# Conflicts:
#	source/game_sa/Audio/CAudioEngine.cpp
#	source/game_sa/Audio/CAudioEngine.h
#	source/game_sa/CBuoyancy.cpp

+++ source/game_sa/CBuoyancy.h
@@ -72,7 +72,7 @@ public:
     void PreCalcSetup(CPhysical* pEntity, float fBuoyancy);
     void AddSplashParticles(CPhysical* pEntity, CVector vecFrom, CVector vecTo, CVector vecSplashDir, unsigned char bReduceParticleSize);
     void SimpleCalcBuoyancy(CPhysical* pEntity);
-    double SimpleSumBuoyancyData(CVector* vecPointRelativeToSurface, eBuoyancyPointState ePointState);
+    float SimpleSumBuoyancyData(CVector* vecPointRelativeToSurface, eBuoyancyPointState ePointState);
     void FindWaterLevel(CVector const& vecInitialZPos, CVector* outVecOffset, eBuoyancyPointState* outInWaterState);
     void FindWaterLevelNorm(CVector const& vecInitialZPos, CVector* outVecOffset, eBuoyancyPointState* outInWaterState, CVector* outVecNormal);
 };
